Port to the new energy api

+++ src/main/kotlin/io/github/lucaargolo/lifts/utils/ModConfig.kt
@@ -10,16 +10,16 @@ class ModConfig {
     class ElectricLiftConfig(
         platformSpeed: Double,
         platformRange: Int,
-        val energyCapacity: Double
+        val energyCapacity: Long
     ): LiftConfig(platformSpeed, platformRange)
 
     class LiftConfigs(
         val stirlingLift: LiftConfig = LiftConfig(1.0, 16),
-        val electricLiftMk1: ElectricLiftConfig = ElectricLiftConfig(1.2, 32, 32000.0),
-        val electricLiftMk2: ElectricLiftConfig = ElectricLiftConfig(1.4, 64, 64000.0),
-        val electricLiftMk3: ElectricLiftConfig = ElectricLiftConfig(1.6, 128, 128000.0),
-        val electricLiftMk4: ElectricLiftConfig = ElectricLiftConfig(1.8, 128, 256000.0),
-        val electricLiftMk5: ElectricLiftConfig = ElectricLiftConfig(2.0, 256, 512000.0)
+        val electricLiftMk1: ElectricLiftConfig = ElectricLiftConfig(1.2, 32, 32000),
+        val electricLiftMk2: ElectricLiftConfig = ElectricLiftConfig(1.4, 64, 64000),
+        val electricLiftMk3: ElectricLiftConfig = ElectricLiftConfig(1.6, 128, 128000),
+        val electricLiftMk4: ElectricLiftConfig = ElectricLiftConfig(1.8, 128, 256000),
+        val electricLiftMk5: ElectricLiftConfig = ElectricLiftConfig(2.0, 256, 512000)
     )
 
     val maxFuelTicksStored = 32000
